Catch slow case in PDF tiling.
Sometimes a PDF file can use just a very small amount of a large tile;
rendering the whole tile is a waste of time. Previous attempts to catch
this case have worked for most cases, but can be defeated when the
area needing rendering crosses a tile edge.

Here we improve the formulation to catch exactly what we want.

This implementation was driven by the example file in Sumatra PDFs
bug tracker, for issue 2248, though the actual code change is
different. Thanks to zeniko, nonetheless for this!