Major changes in the entire design.
Initial commit of this and it has several issues that have to be worked through.  Lots of
broken stuff but I needed to get it in place.

Added 2 new classes.  One is muctx which contains all the interfacing to mupdf and uses
standard c++ types.  The other is mudocument which interfaces between muctx and the WinRT
world.  It passes WinRT objects to and from applications which can be written in C++, C#,
Visual Basic, Javascript etc.   mudocument and muctx will reside in a winRT DLL.   The
Viewer application (c++ version) is now in mupdf_cpp project and now has a flip view
structure that includes a scroll viewer which has data binding to a collection.  This
enables much easier zooming and updating, however I am still trying how to bind the
canvas elements that contain the text search or link information.