Rework of background task set up for rendering the thumbnails of the pages in the background.
these are stored and swapped in and out of the page collection that is viewed as the user goes through
the flip view.  had some issues with respect to getting the background tasks shut down and mupdf
cleaned up when someone opens a new file. that is now all resolved.   This all simplified the code
significantly.