PDF Pattern gstate fix.
The PDF Reference manual is very confusing about what gstate should
be used to run patterns in.

Essentially, my experiments seem to suggest that as we run through a
PDF page, at the start of executing each stream, we should remember
the current gstate as the 'parent' gstate. Then whenever we instantiate
a pattern (via scn etc), we should set the pattern to remember that
gstate. When we come to render the pattern, the pattern should be
rendered using the remembered gstate, not the current one.

This causes many progressions in our tests.