Bug 694092: Android: Redraw performance (overdraw) fixes
Thanks to Goncalo Ferreira(*) (aka monxalo) for this patch.

Firstly, we move our textured background off the layout and into a
style applied to MuPDFActivity. By using "windowBackground", we
avoid the default background being redrawn only to be overlaid
with ours. This cuts out one level of overdrawing.

Secondly, when drawing each PageView, the old code would render the
background for the page, then would draw the bitmap over the top of
that. While it's important to draw the background of the page before
we have a bitmap for the page, we can avoid that stage once a page
bitmap arrives.

(* apologies for not being able to put the cedilla on the 'c' in your
name, but git gives problems with top bit set chars.)