Exception handling changes
In preparation for work on progressive loading, update the exception
handling scheme slightly.

Until now, exceptions (as thrown with fz_throw, and caught with
fz_try/fz_catch) have merely had an informative string. They have
never had anything that can be compared to see if an error is of
a particular type.

We now introduce error codes; when we fz_throw, we now always
give an error code, and can optionally (using fz_throw_message)
give both an error code and an informative string.

When we fz_rethrow from within a fz_catch, both the error code and
the error message is maintained. Using fz_rethrow_message we can
'improve' the error message, but the code is maintained.

The error message can be read out using fz_caught_message() and the
error code can be read as fz_caught().

Currently we only define a 'generic' error. This will expand in future
versions to include other error types that may be tested for.