Initial commit of work towards getting windows phone viewer working
Currently with this commit the calls to the winRT API are working
and I am rendering and viewing pages with a UI element called a LongListSelector.
This is similar to the continuous scroll view.   Also, the manipulations
are working but I have to launch a rerender of the page still.

Launching from file type occurs as well as browsing of the applications
local folder.

Still much to do, including search, contents, hyperlinks, scaling rerender,
a page based view and interface to skydrive.