Fix memory leak due to softmasks.
A user (sebblonline) reports a problem when rendering the 300Meg
document downloadable from:

 http://www.mitsubishicarbide.com/EU/de/product/epaper/index.html

(7th icon on the bottom). Memento builds report that softmask objects
are leaked.

Tracing these it appears that the handling of softmasks in
pdf_run_xobject is not quite nested right. Fixing this reveals another
problem with clipping paths being doubly removed. Adding a gsave/grestore
pair solves this and leaves us with a well behaved program.