SVG: Fix fts_01_0106.pdf
A gradient fill that doesn't fill the bbox should be see through
(unless background color is set, but we'll worry about that case
when I find an example file that uses it). Arrange for the pixmap
we draw the gradient fill into to be transparent initially.

Also ensure that when we convert to png we preserve transparency.