Handle stroke+fill operations with transparency/blendmodes.
When stroking and filling in a single operation, we are supposed
to form the complete stroke+fill image, then blend it back, rather
than filling and blending, then stroking and blending.

This only matters during transparency, or with non-normal blend
modes.

We fix MuPDF to push a knockout group when doing such operations.