Update mupdf to use latest openjpeg pulled from ghostpdl.
The OpenJPEG in gs is v2, with various patches for fixes. These are in
the process of being passed upstream. We now automatically pull the
openjpeg tree out of GhostPDL and put it in as one particular branch
in the thirdparty/openjpeg.git repo. Change to track this in MuPDF.

This is in keeping with what we have been doing with the jbig2dec
repo for a while now.