Fix so that multi-threaded rendering of display list works.
Proper use of mutex lock in areas where we multiple threads can
not be accessing the document or page level objects at the same
time.   With this fix, multiple threads can be rendering different
display lists at the same time.

Also fix for use of page selection scroll bar so that it does not
render the page until the user lifts off the bar.  This is the
same as the Android app and avoids smaller devices getting overloaded
if you do rapid scrolling in a document.