Addition of printing support.
This uses the xaml printing methods.  It is likely that I need to go do a DirectX approach since the xaml
printing method could run into memory problems as it seems the architecture requires all the pages to
be rendered prior to the print job even starting.  Currently checking on this with a contact at MS.

I wanted to get this code in place now since it is working and it has the needed framework for much of
what we want.   A change to DirectX requires a changes in the solution file that may be complicated and so it
makes sense to do a split here with respect to providing printing support.