DirectX printing added
After much research it was determined that use of the DirectX 2D methods is the proper way
to add printing support to a Windows 8 application.   This enables the pages to be rendered
as they are requested by the print thread.   This occurs for both the print preview screen
as well as the print to the output device.