Bug 694116: Solve valgrind issues in draw device.
The actual issue here is that a pixmap is dropped more times than
it should be due to an error in the rendering pipeline.

The problem arises because we fail to push a clip image mask, but
still pop the mask off the stack later. This puts us off by 1 in
the stack handling.

The simplest solution to this (that will be safe no matter what
mistakes are made by the caller too) is to add some simple tests
in the draw device to ensure we do not free too early.