Bug 691691: Add way of clearing cached objects out of the xref.
We add various facilities here, intended to allow us to efficiently
minimise the memory we use for holding cached pdf objects.

Firstly, we add the ability to 'mark' all the currently loaded objects.

Next we add the ability to 'clear the xref' - to drop all the currently
loaded objects that have no other references except the ones held by the
xref table itself.

Finally, we add the ability to 'clear the xref to the last mark' - to
drop all the currently loaded objects that have been created since the
last 'mark' operation and have no other references except the ones held
by the xref table.

We expose this to the user by adding a new device hint 'FZ_NO_CACHE'.
If set on the device, then the PDF interpreter will pdf_mark_xref before
starting and pdf_clear_xref_to_mark afterwards. Thus no additional
objects will be retained in memory after a given page is run, unless
someone else picks them up and takes a reference to them as part of
the run.

We amend our simple example app to set this device hint when loading
pages as part of a search.