Add tiling into the DirectX printing code.
The tiling in x and y is needed to ensure that we can print at high
resolutions with devices that have smaller bit map sizes (e.g. the surface).
Banding only in the y dimension like we often do is not sufficient.

Also fix an open with file association bug that must of occurred with the
transition to 8.1

And update WinRT solution for recent changes in mupdf code.   This includes
the addition of a few new files and the document type registration.