Bug 694897: Fix valgrind issues with versions
If the /Version is a single character string (say "s") then the
current code for converting this in pdf_init_document reads off
the end of the string.

Simple fix is to use fz_atof instead.

Same fix for reading the PDF version normally.

This solves:

  53b830f849d028fb2d528520716e157a_asan_heap-oob_478692_5259_4534.pdf

Thanks to Mateusz Jurczyk and Gynvael Coldwind of the Google Security
Team for providing the example files.