Bug 695021: Fix pdf_insert_page operation with empty page tree.
Patch from Thomas Fach-Pedersen to fix the operation of pdf_insert_page
when called with an empty page tree. Many thanks! As noted in the code
with a FIXME this currently throws an error.

Also, cope with being told to add a page "at" INT_MAX as meaning to
add it at the end of the document.

Possibly this code should cope with a Root without a Pages entry, or
a Pages without a Kids too, but we can fix this in future if it ever
becomes a problem.