Significant amount of content needed for gsview
This includes the interface to ghostscript for creating pdf files from ps
for mupdf to then render in the viewer as well as the capability to print
the output of xpswrite through the windows print queue.   Added thumb
viewing and navigation as well as navigation through the table of contents.

To do.  Add in password handling, Text Search, Hyperlinks, GS conversion to
other file formats, MuPDF save as other file formats, Page extractions,
proper page selection for printing, Zooming (rescale), Fix the page ranges
that are displayed at full resolution during navigation.