Add pdf_process interface.
Currently the only processing we can do of PDF pages is to run
them through an fz_device. We introduce new "pdf_process"
functionality here to enable us to do more things.

We define a pdf_processor structure with a set of function
pointers in, one per PDF operator, together with functions
for processing xobjects etc. The guts of pdf_run_page_contents
and pdf_run_annot operations are then extracted to give
pdf_process_page_contents and pdf_process_annot, and the
originals implemented in terms of these.

This commit contains just one instance of a pdf_processor, namely
the "run" processor, which contains the original code refactored.

The graphical state (and device pointer) is now part of private data
to the run operator set, rather than being in pdf_csi.