Ensure that BDC operators get both params.
Currently, when parsing, each time we encounter a name, we throw away
the last name we had. BDC operators are called with:

 /Name <object> BDC

If the <object> is a name, we lose the original /Name.

To fix this, parsing a name when we already have a name will cause
the name to be stored as an object.

This has various knock on effects throughout the code to read from
csi->obj rather than csi->name.

Also, ensure that when cleaning, we collect a list of the object
names in our new resources dictionary.